fix: CWE-20

+++ app.py
@@ -74,9 +74,9 @@ def update_account_balance(username, operation, amount):
         balance = cur.fetchone()[0]
         # print("1", balance)
         if operation == 'withdrawal':
-            balance -= int(amount)
+            balance -= float(amount)
         elif operation == 'deposit':
-            balance += int(amount)
+            balance += float(amount)
         if balance < 0:
 
             response = "Failed to process your request, please check the amount you entered and retry later."
